Split up InvocationTracePanel
