+++ packages/devtools/devtools/src/panels/edge/InvocationTracePanel/index.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2020 DXOS.org
+// Copyright 2025 DXOS.org
 //
 
 import { InvocationTracePanel } from './InvocationTracePanel';
