+++ packages/devtools/devtools/src/panels/edge/InvocationTracePanel/utils.ts
@@ -1,0 +1,7 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+export const formatDuration = (duration: number): string => {
+  return `${(duration / 1000).toFixed(2)}`;
+};
